requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# Google AI
google-genai>=0.1.0
google-generativeai>=0.8.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Visualization (optional)
ipython>=8.0.0
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
